[SPARK-37924][SQL] Sort table properties by key in SHOW CREATE TABLE on VIEW (v1)
### What changes were proposed in this pull request?

This PR is a sort of a followup of https://github.com/apache/spark/pull/34719. It added a test but it is flaky due to the order of `TABLPROPERTIES` in `SHOW CREATE TALBE` on `VIEW` in v1 code path.

This PR proposes to have a deterministic order by sorting the table properties in the show command. This is already being sorted in v2 (see `ShowCreateTableExec`).

### Why are the changes needed?

To have the deterministic order, and fix the flaky test.

### Does this PR introduce _any_ user-facing change?

Virtually no. It might affect the order of TABLEPROPERTIES in `SHOW TABLE`'s output on `VIEW` when users are rely on

### How was this patch tested?

Fixed the flaky unittest to explicitly test the order.

Closes #35222 from HyukjinKwon/SPARK-37924.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>