[SPARK-37949][SQL] Improve Rebalance statistics estimation
### What changes were proposed in this pull request?

Match `RebalancePartitions` in `SizeInBytesOnlyStatsPlanVisitor` and `BasicStatsPlanVisitor`.

### Why are the changes needed?

The defualt statistics estimation only consider the size in bytes, which may lost the row rount and columns statistics.

The `RebalancePartitions` actually does not change the statistics of plan, so we can use the statistics of its child for more accurate.

### Does this PR introduce _any_ user-facing change?

no, only affect the statistics of plan

### How was this patch tested?

Unify the test in `BasicStatsEstimationSuite`

Closes #35235 from ulysses-you/SPARK-37949.

Authored-by: ulysses-you <ulyssesyou18@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>