[SPARK-38131][SQL] Use error classes in user-facing exceptions only
### What changes were proposed in this pull request?
In the PR, I propose to remove the error class `ROW_FROM_CSV_PARSER_NOT_EXPECTED` and don't use the error class `UNSUPPORTED_FEATURE` in an up-cast exception which is an internal one.

### Why are the changes needed?
The error classes are supposed to be used in user-facing errors/exceptions only. It doesn't make sense to introduce/use them in internal errors.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
By running existing test suites.

Closes #35445 from MaxGekk/keep-only-user-facing-error-classes.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>