[SPARK-38178][SS] Correct the logic to measure the memory usage of RocksDB
### What changes were proposed in this pull request?

Correct the logic to measure the memory usage of RocksDB to include the memory used by block cache.
As "block-cache-pinned-usage" is included in "block-cache-usage", we don't need to sum the pinned usage separately.

### Why are the changes needed?

Current reported metrics of RocksDB memory usage is not correct.

### Does this PR introduce _any_ user-facing change?

NO

### How was this patch tested?

The information could refer to https://github.com/facebook/rocksdb/wiki/Memory-usage-in-RocksDB

Closes #35480 from Myasuka/rocksdb-mem-usage.

Authored-by: Yun Tang <myasuka@live.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>