[SPARK-38192][CORE][TESTS] Use `try-with-resources` in `Level/RocksDBSuite.java`
### What changes were proposed in this pull request?
`countKeys` method in `LevelDBSuite` use `db.db().newIterator()` to create a `DBIterator` instance, but there is no way to close it because the iterator not registered to `LevelDB.iteratorTracker`, and the similar issue also exists in `RocksDBSuite`, so  this pr use `try-with-resources` to ensure the iterator closed.

### Why are the changes needed?
Bug fix.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Pass GA

Closes #35493 from LuciferYang/SPARK-38192.

Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>