[MINOR][DOC] Fix documentation for structured streaming - addListener
### What changes were proposed in this pull request?

This PR fixes the incorrect documentation in Structured Streaming Guide where it says `sparkSession.streams.attachListener()` instead of `sparkSession.streams.addListener()` which is the correct usage as mentioned in the code snippet below in the same doc.

![image](https://user-images.githubusercontent.com/298735/154593814-0a865311-b168-4929-b4af-a8c939168f26.png)

### Why are the changes needed?

The documentation was erroneous, and needs to be fixed to avoid confusion by readers

### Does this PR introduce _any_ user-facing change?

Yes, since it's a doc fix. This fix needs to be applied to previous versions retro-actively as well.

### How was this patch tested?

Not necessary

Closes #35562 from yeskarthik/fix-structured-streaming-docs-1.

Authored-by: Karthik Subramanian <karsubr@microsoft.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>