[SPARK-38285][SQL] Avoid generator pruning for invalid extractor
### What changes were proposed in this pull request?

This fixes a bug in generator nested column pruning. The bug happens when the extractor pattern is like `GetArrayStructFields(GetStructField(...), ...)` on the generator output. Once the input to the generator is an array, after replacing with the extractor based on pruning logic, it becomes an extractor of `GetArrayStructFields(GetArrayStructFields(...), ...)` which is not valid.

### Why are the changes needed?

To fix a bug in generator nested column pruning.

### Does this PR introduce _any_ user-facing change?

Yes, fixing a user-facing bug.

### How was this patch tested?

Added unit test.

Closes #35749 from viirya/SPARK-38285.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
Signed-off-by: Liang-Chi Hsieh <viirya@gmail.com>