[SPARK-38240][SQL][FOLLOW-UP] Make RuntimeReplaceableAggregate as an add-on of AggregateFunction
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/35534 (https://github.com/apache/spark/pull/35534#discussion_r809708015) that proposes to make `RuntimeReplaceableAggregate` as an add-on of `AggregateFunction`.

Now, `ReplaceableAggregateFunction` is a mix-in for `AggregateFunction` that makes the expression can be replaced runtime.

### Why are the changes needed?

To make the hierarchy of class similar.

### Does this PR introduce _any_ user-facing change?

No, dev-only.

### How was this patch tested?

I tested that it compiles fine. CI in this PR will test it out.

Closes #35626 from HyukjinKwon/SPARK-38240-followup.

Lead-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>