[SPARK-38423][K8S][FOLLOWUP] PodGroup spec should not be null
### What changes were proposed in this pull request?

This PR is a follow-up of #35639 to prevent `PodGroup` with `null` spec.

### Why are the changes needed?

First, for `null` value, we should use `Option(null)` to make it `None`. This PR fixes the following.
```
- priorityClassName = Some(pod.pod.getSpec.getPriorityClassName)
+ priorityClassName = Option(pod.pod.getSpec.getPriorityClassName)
```

Second, when `queue` and `priorityClassName` are not given. `PodGroup` is created with `null` spec.
This causes NullPointerException during the test case. This PR fixes it.

### Does this PR introduce _any_ user-facing change?

No. This is not released yet.

### How was this patch tested?

Pass the CIs.

Closes #35769 from dongjoon-hyun/SPARK-38423.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>