[SPARK-38443][SS][DOC] Document config STREAMING_SESSION_WINDOW_MERGE_SESSIONS_IN_LOCAL_PARTITION
### What changes were proposed in this pull request?

This proposes to document the config `spark.sql.streaming.sessionWindow.merge.sessions.in.local.partition` in structured streaming guide.

### Why are the changes needed?

We have use case the customer faces issue on large shuffle write in session window. Users are hardly to know what is the issue and cannot find related document. The config is hidden, although it is useful for such case. We should document it so end users can find it easily.

### Does this PR introduce _any_ user-facing change?

Yes, documented a config in user-facing guide.

### How was this patch tested?

Basically doc change.

Closes #35762 from viirya/ss_doc.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
Signed-off-by: Liang-Chi Hsieh <viirya@gmail.com>