[SPARK-38538][K8S][TESTS] Fix driver environment verification in BasicDriverFeatureStepSuite
### What changes were proposed in this pull request?

This PR aims to fix the driver environment verification logic in `BasicDriverFeatureStepSuite`.

### Why are the changes needed?

When SPARK-25876 added a test logic at Apache Spark 3.0.0, it used `envs(v) === v` instead of `envs(k) === v`.
https://github.com/apache/spark/blob/923b2f18cc9c1bd6a3aea0a1fb7eff802199d2c2/resource-managers/kubernetes/core/src/test/scala/org/apache/spark/deploy/k8s/features/BasicDriverFeatureStepSuite.scala#L94-L96

This bug was hidden because the test key-value pairs have identical set. If we have different strings for keys and values, the test case fails.

https://github.com/apache/spark/blob/923b2f18cc9c1bd6a3aea0a1fb7eff802199d2c2/resource-managers/kubernetes/core/src/test/scala/org/apache/spark/deploy/k8s/features/BasicDriverFeatureStepSuite.scala#L42-L44

### Does this PR introduce _any_ user-facing change?

To have a correct test coverage.

### How was this patch tested?

Pass the CIs.

Closes #35828 from dongjoon-hyun/SPARK-38538.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Yuming Wang <yumwang@ebay.com>