[SPARK-38593][SS] Carry over the metric of the number of dropped late events into SessionWindowStateStoreSaveExec
### What changes were proposed in this pull request?

This PR proposes to carry over the metric of the number of dropped late events from SessionWindowStateStoreRestoreExec to SessionWindowStateStoreSaveExec, so that state operator progress for session window aggregation operator will have correct number.

### Why are the changes needed?

The number of dropped late events for session window aggregation was missed in the mechanism of SS metrics (state operator progress).

### Does this PR introduce _any_ user-facing change?

Yes, the number of dropped late events for session window aggregation will be taken into account for state operator progress in streaming listener, as well as a graph on late events in the SS UI page.

### How was this patch tested?

Modified UT.

Closes #35909 from HeartSaVioR/SPARK-38593.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>