[SPARK-38548][SQL] New SQL function: try_sum
### What changes were proposed in this pull request?

Add a new SQL function: try_sum. It is identical to the function `sum`, except that it returns `NULL` result instead of throwing an exception on integral/decimal value overflow.
Note it is different from sum when ANSI mode is off:
| Function         | Sum                                | TrySum      |
|------------------|------------------------------------|-------------|
| Decimal overflow | Return NULL                        | Return NULL |
| Integer overflow | Return lower 64 bits of the result | Return NULL |

### Why are the changes needed?

* Users can manage to finish queries without interruptions in ANSI mode.
* Users can get NULLs instead of unreasonable results if overflow occurs when ANSI mode is off. For example
```
> SELECT sum(col) FROM VALUES (9223372036854775807L), (1L) AS tab(col);
-9223372036854775808

> SELECT try_sum(col) FROM VALUES (9223372036854775807L), (1L) AS tab(col);
NULL
```

### Does this PR introduce _any_ user-facing change?

Yes, a new SQL function: try_sum which is identical to the function `sum`, except that it returns `NULL` result instead of throwing an exception on integral/decimal value overflow.

### How was this patch tested?

UT

Closes #35848 from gengliangwang/trySum2.

Authored-by: Gengliang Wang <gengliang@apache.org>
Signed-off-by: Gengliang Wang <gengliang@apache.org>