[SPARK-38604][SQL] Keep ceil and floor with only a single argument the same as before
This is just the fix. I didn't add any tests yet. I am happy to do it, I just wasn't sure where the right place to put the tests would be. Once I have tests I will cherry-pick this back to the 3.3 branch and put up a PR for that too. I am also happy to update the comments because this is a bit confusing that there is no indication that things have changed.

Closes #35913 from revans2/ceil_floor_no_arg_behavior.

Authored-by: Robert (Bobby) Evans <bobby@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>